config.yml: gems was deprecated in favour of plugins
The gems key in config.yml files was replaced by plugins on Jekyll 3.5.
While it only triggers a Deprecation warning, the warning is visible
during site compilation. It's probably a good idea to update the docs
using the new keys.

The test cases have been updated as well to use the new keys on their
config.yml files so that deprecation warnings aren't logged to the
terminal as well.